fix: change DELETE in blogs-router.ts & posts-router.ts

+++ src/routes/posts-router.ts
@@ -60,7 +60,7 @@ export const getPostsRouter = () => {
 
     router.delete('/:id', authMiddleware, async (req: Request, res: Response) => {
         const postForDelete = postsRepository.findPostById(req.params.id)
-        if (!postForDelete) return res.sendStatus(HTTP_STATUS.NOT_FOUND_404)
+        if (!postForDelete || null) return res.sendStatus(HTTP_STATUS.NOT_FOUND_404)
 
         await postsRepository.deletePostById(req.params.id)
         res.sendStatus(HTTP_STATUS.NO_CONTENT_204)
